objectstate: rte: move mcp-specific code
previously we didn't have a machineconfig specific file
to hold all the code which depends on the MCO API, so it
was scattered in rte.go. But now we have, so let's isolate
the API-dependent code from the API-agnostic code further.

Internal code movement (+ trivial import fix), no change
in behavior.
